rename from `services` to `modules`
