+++ .dev.vars.example
@@ -6,3 +6,6 @@ AI_GATEWAY_BASE_URL=https://gateway.ai.cloudflare.com/v1/YOUR_ACCOUNT_ID/YOUR_GA
 
 # Cloudflare AI Gateway Token (optional)
 CF_AIG_TOKEN=your-cf-aig-token-here
+
+# Database (Connection String)
+DATABASE_URL=postgres://username:password@host:port/database_name
