+++ src/lib/modules/paceNote/__tests__/workers-ai.service.test.ts
