+++ src/lib/modules/paceNote/prompts/base.md
