+++ src/lib/modules/policyFoo/doadFoo/metadata-selector.ts
