+++ src/lib/modules/policyFoo/doadFoo/prompts/main.md
