+++ src/lib/modules/policyFoo/doadFoo/prompts/metadata-selector.md
